Tests to check pipeline restartability.
We check pipeline restarting both without/with bloocoo error correction.
Restarting the pipeline is needed if the minia loop is interrupted on error (job limitation in time, memory issue, etc.), or to resume iterations.

This is to solve a pipeline bug pointed out by S. Letort : restarting doesn't work with bloocoo.

[redmine issue 54]